requirements.txt
python-dotenv
datasets
apscheduler
PyGithub
firebase-admin
fastapi
uvicorn
requests
pinecone
beautifulsoup4
feedparser
neo4j
torch
transformers
sentence-transformers
scikit-learn
stable-baselines3[extra]
gym
docker
psutil
chromadb